Unverted imgproc-geometry dependency and moved more functions to geometry.
