+++ modules/geometry/include/opencv2/geometry/3d.hpp
@@ -1280,25 +1280,6 @@ CV_EXPORTS_W int solvePnPGeneric( InputArray objectPoints, InputArray imagePoint
                                   InputArray rvec = noArray(), InputArray tvec = noArray(),
                                   OutputArray reprojectionError = noArray() );
 
-/** @brief Draw axes of the world/object coordinate system from pose estimation. @sa solvePnP
-
-@param image Input/output image. It must have 1 or 3 channels. The number of channels is not altered.
-@param cameraMatrix Input 3x3 floating-point matrix of camera intrinsic parameters.
-\f$\cameramatrix{A}\f$
-@param distCoeffs Input vector of distortion coefficients
-\f$\distcoeffs\f$. If the vector is empty, the zero distortion coefficients are assumed.
-@param rvec Rotation vector (see @ref Rodrigues ) that, together with tvec, brings points from
-the model coordinate system to the camera coordinate system.
-@param tvec Translation vector.
-@param length Length of the painted axes in the same unit than tvec (usually in meters).
-@param thickness Line thickness of the painted axes.
-
-This function draws the axes of the world/object coordinate system w.r.t. to the camera frame.
-OX is drawn in red, OY in green and OZ in blue.
- */
-CV_EXPORTS_W void drawFrameAxes(InputOutputArray image, InputArray cameraMatrix, InputArray distCoeffs,
-                                InputArray rvec, InputArray tvec, float length, int thickness=3);
-
 /** @brief Converts points from Euclidean to homogeneous space.
 
 @param src Input vector of N-dimensional points.
@@ -2216,201 +2197,6 @@ CV_EXPORTS_W void filterHomographyDecompByVisibleRefpoints(InputArrayOfArrays ro
                                                            OutputArray possibleSolutions,
                                                            InputArray pointsMask = noArray());
 
-//! cv::undistort mode
-enum UndistortTypes
-{
-    PROJ_SPHERICAL_ORTHO  = 0,
-    PROJ_SPHERICAL_EQRECT = 1
-};
-
-/** @brief Transforms an image to compensate for lens distortion.
-
-The function transforms an image to compensate radial and tangential lens distortion.
-
-The function is simply a combination of #initUndistortRectifyMap (with unity R ) and #remap
-(with bilinear interpolation). See the former function for details of the transformation being
-performed.
-
-Those pixels in the destination image, for which there is no correspondent pixels in the source
-image, are filled with zeros (black color).
-
-A particular subset of the source image that will be visible in the corrected image can be regulated
-by newCameraMatrix. You can use #getOptimalNewCameraMatrix to compute the appropriate
-newCameraMatrix depending on your requirements.
-
-The camera matrix and the distortion parameters can be determined using #calibrateCamera. If
-the resolution of images is different from the resolution used at the calibration stage, \f$f_x,
-f_y, c_x\f$ and \f$c_y\f$ need to be scaled accordingly, while the distortion coefficients remain
-the same.
-
-@param src Input (distorted) image.
-@param dst Output (corrected) image that has the same size and type as src .
-@param cameraMatrix Input camera matrix \f$A = \vecthreethree{f_x}{0}{c_x}{0}{f_y}{c_y}{0}{0}{1}\f$ .
-@param distCoeffs Input vector of distortion coefficients
-\f$(k_1, k_2, p_1, p_2[, k_3[, k_4, k_5, k_6[, s_1, s_2, s_3, s_4[, \tau_x, \tau_y]]]])\f$
-of 4, 5, 8, 12 or 14 elements. If the vector is NULL/empty, the zero distortion coefficients are assumed.
-@param newCameraMatrix Camera matrix of the distorted image. By default, it is the same as
-cameraMatrix but you may additionally scale and shift the result by using a different matrix.
- */
-CV_EXPORTS_W void undistort( InputArray src, OutputArray dst,
-                             InputArray cameraMatrix,
-                             InputArray distCoeffs,
-                             InputArray newCameraMatrix = noArray() );
-
-/** @brief Computes the undistortion and rectification transformation map.
-
-The function computes the joint undistortion and rectification transformation and represents the
-result in the form of maps for #remap. The undistorted image looks like original, as if it is
-captured with a camera using the camera matrix =newCameraMatrix and zero distortion. In case of a
-monocular camera, newCameraMatrix is usually equal to cameraMatrix, or it can be computed by
-#getOptimalNewCameraMatrix for a better control over scaling. In case of a stereo camera,
-newCameraMatrix is normally set to P1 or P2 computed by #stereoRectify .
-
-Also, this new camera is oriented differently in the coordinate space, according to R. That, for
-example, helps to align two heads of a stereo camera so that the epipolar lines on both images
-become horizontal and have the same y- coordinate (in case of a horizontally aligned stereo camera).
-
-The function actually builds the maps for the inverse mapping algorithm that is used by #remap. That
-is, for each pixel \f$(u, v)\f$ in the destination (corrected and rectified) image, the function
-computes the corresponding coordinates in the source image (that is, in the original image from
-camera). The following process is applied:
-\f[
-\begin{array}{l}
-x  \leftarrow (u - {c'}_x)/{f'}_x  \\
-y  \leftarrow (v - {c'}_y)/{f'}_y  \\
-{[X\,Y\,W]} ^T  \leftarrow R^{-1}*[x \, y \, 1]^T  \\
-x'  \leftarrow X/W  \\
-y'  \leftarrow Y/W  \\
-r^2  \leftarrow x'^2 + y'^2 \\
-x''  \leftarrow x' \frac{1 + k_1 r^2 + k_2 r^4 + k_3 r^6}{1 + k_4 r^2 + k_5 r^4 + k_6 r^6}
-+ 2p_1 x' y' + p_2(r^2 + 2 x'^2)  + s_1 r^2 + s_2 r^4\\
-y''  \leftarrow y' \frac{1 + k_1 r^2 + k_2 r^4 + k_3 r^6}{1 + k_4 r^2 + k_5 r^4 + k_6 r^6}
-+ p_1 (r^2 + 2 y'^2) + 2 p_2 x' y' + s_3 r^2 + s_4 r^4 \\
-s\vecthree{x'''}{y'''}{1} =
-\vecthreethree{R_{33}(\tau_x, \tau_y)}{0}{-R_{13}((\tau_x, \tau_y)}
-{0}{R_{33}(\tau_x, \tau_y)}{-R_{23}(\tau_x, \tau_y)}
-{0}{0}{1} R(\tau_x, \tau_y) \vecthree{x''}{y''}{1}\\
-map_x(u,v)  \leftarrow x''' f_x + c_x  \\
-map_y(u,v)  \leftarrow y''' f_y + c_y
-\end{array}
-\f]
-where \f$(k_1, k_2, p_1, p_2[, k_3[, k_4, k_5, k_6[, s_1, s_2, s_3, s_4[, \tau_x, \tau_y]]]])\f$
-are the distortion coefficients.
-
-In case of a stereo camera, this function is called twice: once for each camera head, after
-#stereoRectify, which in its turn is called after #stereoCalibrate. But if the stereo camera
-was not calibrated, it is still possible to compute the rectification transformations directly from
-the fundamental matrix using #stereoRectifyUncalibrated. For each camera, the function computes
-homography H as the rectification transformation in a pixel domain, not a rotation matrix R in 3D
-space. R can be computed from H as
-\f[\texttt{R} = \texttt{cameraMatrix} ^{-1} \cdot \texttt{H} \cdot \texttt{cameraMatrix}\f]
-where cameraMatrix can be chosen arbitrarily.
-
-@param cameraMatrix Input camera matrix \f$A=\vecthreethree{f_x}{0}{c_x}{0}{f_y}{c_y}{0}{0}{1}\f$ .
-@param distCoeffs Input vector of distortion coefficients
-\f$(k_1, k_2, p_1, p_2[, k_3[, k_4, k_5, k_6[, s_1, s_2, s_3, s_4[, \tau_x, \tau_y]]]])\f$
-of 4, 5, 8, 12 or 14 elements. If the vector is NULL/empty, the zero distortion coefficients are assumed.
-@param R Optional rectification transformation in the object space (3x3 matrix). R1 or R2 ,
-computed by #stereoRectify can be passed here. If the matrix is empty, the identity transformation
-is assumed. In #initUndistortRectifyMap R assumed to be an identity matrix.
-@param newCameraMatrix New camera matrix \f$A'=\vecthreethree{f_x'}{0}{c_x'}{0}{f_y'}{c_y'}{0}{0}{1}\f$.
-@param size Undistorted image size.
-@param m1type Type of the first output map that can be CV_32FC1, CV_32FC2 or CV_16SC2, see #convertMaps
-@param map1 The first output map.
-@param map2 The second output map.
- */
-CV_EXPORTS_W
-void initUndistortRectifyMap(InputArray cameraMatrix, InputArray distCoeffs,
-                             InputArray R, InputArray newCameraMatrix,
-                             Size size, int m1type, OutputArray map1, OutputArray map2);
-
-/** @brief Computes the projection and inverse-rectification transformation map. In essense, this is the inverse of
-#initUndistortRectifyMap to accomodate stereo-rectification of projectors ('inverse-cameras') in projector-camera pairs.
-
-The function computes the joint projection and inverse rectification transformation and represents the
-result in the form of maps for #remap. The projected image looks like a distorted version of the original which,
-once projected by a projector, should visually match the original. In case of a monocular camera, newCameraMatrix
-is usually equal to cameraMatrix, or it can be computed by
-#getOptimalNewCameraMatrix for a better control over scaling. In case of a projector-camera pair,
-newCameraMatrix is normally set to P1 or P2 computed by #stereoRectify .
-
-The projector is oriented differently in the coordinate space, according to R. In case of projector-camera pairs,
-this helps align the projector (in the same manner as #initUndistortRectifyMap for the camera) to create a stereo-rectified pair. This
-allows epipolar lines on both images to become horizontal and have the same y-coordinate (in case of a horizontally aligned projector-camera pair).
-
-The function builds the maps for the inverse mapping algorithm that is used by #remap. That
-is, for each pixel \f$(u, v)\f$ in the destination (projected and inverse-rectified) image, the function
-computes the corresponding coordinates in the source image (that is, in the original digital image). The following process is applied:
-
-\f[
-\begin{array}{l}
-\text{newCameraMatrix}\\
-x  \leftarrow (u - {c'}_x)/{f'}_x  \\
-y  \leftarrow (v - {c'}_y)/{f'}_y  \\
-
-\\\text{Undistortion}
-\\\scriptsize{\textit{though equation shown is for radial undistortion, function implements cv::undistortPoints()}}\\
-r^2  \leftarrow x^2 + y^2 \\
-\theta \leftarrow \frac{1 + k_1 r^2 + k_2 r^4 + k_3 r^6}{1 + k_4 r^2 + k_5 r^4 + k_6 r^6}\\
-x' \leftarrow \frac{x}{\theta} \\
-y'  \leftarrow \frac{y}{\theta} \\
-
-\\\text{Rectification}\\
-{[X\,Y\,W]} ^T  \leftarrow R*[x' \, y' \, 1]^T  \\
-x''  \leftarrow X/W  \\
-y''  \leftarrow Y/W  \\
-
-\\\text{cameraMatrix}\\
-map_x(u,v)  \leftarrow x'' f_x + c_x  \\
-map_y(u,v)  \leftarrow y'' f_y + c_y
-\end{array}
-\f]
-where \f$(k_1, k_2, p_1, p_2[, k_3[, k_4, k_5, k_6[, s_1, s_2, s_3, s_4[, \tau_x, \tau_y]]]])\f$
-are the distortion coefficients vector distCoeffs.
-
-In case of a stereo-rectified projector-camera pair, this function is called for the projector while #initUndistortRectifyMap is called for the camera head.
-This is done after #stereoRectify, which in turn is called after #stereoCalibrate. If the projector-camera pair
-is not calibrated, it is still possible to compute the rectification transformations directly from
-the fundamental matrix using #stereoRectifyUncalibrated. For the projector and camera, the function computes
-homography H as the rectification transformation in a pixel domain, not a rotation matrix R in 3D
-space. R can be computed from H as
-\f[\texttt{R} = \texttt{cameraMatrix} ^{-1} \cdot \texttt{H} \cdot \texttt{cameraMatrix}\f]
-where cameraMatrix can be chosen arbitrarily.
-
-@param cameraMatrix Input camera matrix \f$A=\vecthreethree{f_x}{0}{c_x}{0}{f_y}{c_y}{0}{0}{1}\f$ .
-@param distCoeffs Input vector of distortion coefficients
-\f$(k_1, k_2, p_1, p_2[, k_3[, k_4, k_5, k_6[, s_1, s_2, s_3, s_4[, \tau_x, \tau_y]]]])\f$
-of 4, 5, 8, 12 or 14 elements. If the vector is NULL/empty, the zero distortion coefficients are assumed.
-@param R Optional rectification transformation in the object space (3x3 matrix). R1 or R2,
-computed by #stereoRectify can be passed here. If the matrix is empty, the identity transformation
-is assumed.
-@param newCameraMatrix New camera matrix \f$A'=\vecthreethree{f_x'}{0}{c_x'}{0}{f_y'}{c_y'}{0}{0}{1}\f$.
-@param size Distorted image size.
-@param m1type Type of the first output map. Can be CV_32FC1, CV_32FC2 or CV_16SC2, see #convertMaps
-@param map1 The first output map for #remap.
-@param map2 The second output map for #remap.
- */
-CV_EXPORTS_W
-void initInverseRectificationMap( InputArray cameraMatrix, InputArray distCoeffs,
-                           InputArray R, InputArray newCameraMatrix,
-                           const Size& size, int m1type, OutputArray map1, OutputArray map2 );
-
-//! initializes maps for #remap for wide-angle
-CV_EXPORTS
-float initWideAngleProjMap(InputArray cameraMatrix, InputArray distCoeffs,
-                           Size imageSize, int destImageWidth,
-                           int m1type, OutputArray map1, OutputArray map2,
-                           enum UndistortTypes projType = PROJ_SPHERICAL_EQRECT, double alpha = 0);
-static inline
-float initWideAngleProjMap(InputArray cameraMatrix, InputArray distCoeffs,
-                           Size imageSize, int destImageWidth,
-                           int m1type, OutputArray map1, OutputArray map2,
-                           int projType, double alpha = 0)
-{
-    return initWideAngleProjMap(cameraMatrix, distCoeffs, imageSize, destImageWidth,
-                                m1type, map1, map2, (UndistortTypes)projType, alpha);
-}
-
 /** @brief Computes useful camera characteristics from the camera intrinsic matrix.
  *
  * @param cameraMatrix Input camera intrinsic matrix that can be estimated by #calibrateCamera or
@@ -2701,55 +2487,6 @@ length. Balance is in range of [0, 1].
  */
 CV_EXPORTS_W void estimateNewCameraMatrixForUndistortRectify(InputArray K, InputArray D, const Size &image_size, InputArray R,
     OutputArray P, double balance = 0.0, const Size& new_size = Size(), double fov_scale = 1.0);
-
-/** @brief Computes undistortion and rectification maps for image transform by cv::remap(). If D is empty zero
-distortion is used, if R or P is empty identity matrixes are used.
-
-@param K Camera intrinsic matrix \f$cameramatrix{K}\f$.
-@param D Input vector of distortion coefficients \f$\distcoeffsfisheye\f$.
-@param R Rectification transformation in the object space: 3x3 1-channel, or vector: 3x1/1x3
-1-channel or 1x1 3-channel
-@param P New camera intrinsic matrix (3x3) or new projection matrix (3x4)
-@param size Undistorted image size.
-@param m1type Type of the first output map that can be CV_32FC1 or CV_16SC2 . See convertMaps()
-for details.
-@param map1 The first output map.
-@param map2 The second output map.
- */
-CV_EXPORTS_W void initUndistortRectifyMap(InputArray K, InputArray D, InputArray R, InputArray P,
-    const cv::Size& size, int m1type, OutputArray map1, OutputArray map2);
-
-/** @brief Transforms an image to compensate for fisheye lens distortion.
-
-@param distorted image with fisheye lens distortion.
-@param undistorted Output image with compensated fisheye lens distortion.
-@param K Camera intrinsic matrix \f$cameramatrix{K}\f$.
-@param D Input vector of distortion coefficients \f$\distcoeffsfisheye\f$.
-@param Knew Camera intrinsic matrix of the distorted image. By default, it is the identity matrix but you
-may additionally scale and shift the result by using a different matrix.
-@param new_size the new size
-
-The function transforms an image to compensate radial and tangential lens distortion.
-
-The function is simply a combination of #cv::fisheye::initUndistortRectifyMap (with unity R ) and remap
-(with bilinear interpolation). See the former function for details of the transformation being
-performed.
-
-See below the results of undistortImage.
-   -   a\) result of undistort of perspective camera model (all possible coefficients (k_1, k_2, k_3,
-        k_4, k_5, k_6) of distortion were optimized under calibration)
-    -   b\) result of #cv::fisheye::undistortImage of fisheye camera model (all possible coefficients (k_1, k_2,
-        k_3, k_4) of fisheye distortion were optimized under calibration)
-    -   c\) original image was captured with fisheye lens
-
-Pictures a) and b) almost the same. But if we consider points of image located far from the center
-of image, we can notice that on image a) these points are distorted.
-
-![image](pics/fisheye_undistorted.jpg)
- */
-CV_EXPORTS_W void undistortImage(InputArray distorted, OutputArray undistorted,
-    InputArray K, InputArray D, InputArray Knew = cv::noArray(), const Size& new_size = Size());
-
 /**
 @brief Finds an object pose from 3D-2D point correspondences for fisheye camera model.
