+++ modules/geometry/include/opencv2/geometry/detail/gcgraph.hpp
