+++ modules/geometry/misc/java/gen_dict.json
@@ -17,6 +17,25 @@
                               "dst" : {"ctype" : "vector_Point2f"} },
         "projectPoints" : { "objectPoints" : {"ctype" : "vector_Point3f"},
                             "imagePoints"  : {"ctype" : "vector_Point2f"},
-                            "distCoeffs"   : {"ctype" : "vector_double" } }
+                            "distCoeffs"   : {"ctype" : "vector_double" } },
+        "minEnclosingCircle" : { "points" : {"ctype" : "vector_Point2f"} },
+        "fitEllipse" : { "points" : {"ctype" : "vector_Point2f"} },
+        "fillPoly"   : { "pts" : {"ctype" : "vector_vector_Point"} },
+        "polylines"  : { "pts" : {"ctype" : "vector_vector_Point"} },
+        "fillConvexPoly" : { "points" : {"ctype" : "vector_Point"} },
+        "approxPolyDP" : { "curve"       : {"ctype" : "vector_Point2f"},
+                           "approxCurve" : {"ctype" : "vector_Point2f"} },
+        "arcLength" : { "curve" : {"ctype" : "vector_Point2f"} },
+        "pointPolygonTest" : { "contour" : {"ctype" : "vector_Point2f"} },
+        "minAreaRect" : { "points" : {"ctype" : "vector_Point2f"} },
+        "getAffineTransform" : { "src" : {"ctype" : "vector_Point2f"},
+                                 "dst" : {"ctype" : "vector_Point2f"} },
+        "convexityDefects" : { "contour"          : {"ctype" : "vector_Point"},
+                               "convexhull"       : {"ctype" : "vector_int"},
+                               "convexityDefects" : {"ctype" : "vector_Vec4i"} },
+        "isContourConvex" : { "contour" : {"ctype" : "vector_Point"} },
+        "convexHull" : { "points" : {"ctype" : "vector_Point"},
+                         "hull"   : {"ctype" : "vector_int"},
+                         "returnPoints" : {"ctype" : ""} }
     }
 }
