+++ modules/geometry/src/solvepnp.cpp
@@ -87,47 +87,6 @@ static bool approxEqual(double a, double b, double eps)
 }
 #endif
 
-void drawFrameAxes(InputOutputArray image, InputArray cameraMatrix, InputArray distCoeffs,
-                   InputArray rvec, InputArray tvec, float length, int thickness)
-{
-    CV_INSTRUMENT_REGION();
-
-    int type = image.type();
-    int cn = CV_MAT_CN(type);
-    CV_CheckType(type, cn == 1 || cn == 3 || cn == 4,
-                 "Number of channels must be 1, 3 or 4" );
-
-    cv::Mat img = image.getMat();
-    CV_Assert(img.total() > 0);
-    CV_Assert(length > 0);
-
-    // project axes points
-    std::vector<Point3f> axesPoints;
-    axesPoints.push_back(Point3f(0, 0, 0));
-    axesPoints.push_back(Point3f(length, 0, 0));
-    axesPoints.push_back(Point3f(0, length, 0));
-    axesPoints.push_back(Point3f(0, 0, length));
-    std::vector<Point2f> imagePoints;
-    projectPoints(axesPoints, rvec, tvec, cameraMatrix, distCoeffs, imagePoints);
-
-    cv::Rect imageRect(0, 0, img.cols, img.rows);
-    bool allIn = true;
-    for (size_t i = 0; i < imagePoints.size(); i++)
-    {
-        allIn &= imageRect.contains(imagePoints[i]);
-    }
-
-    if (!allIn)
-    {
-        CV_LOG_WARNING(NULL, "Some of projected axes endpoints are out of frame. The drawn axes may be not reliable.");
-    }
-
-    // draw axes lines
-    line(image, imagePoints[0], imagePoints[1], Scalar(0, 0, 255), thickness);
-    line(image, imagePoints[0], imagePoints[2], Scalar(0, 255, 0), thickness);
-    line(image, imagePoints[0], imagePoints[3], Scalar(255, 0, 0), thickness);
-}
-
 bool solvePnP( InputArray opoints, InputArray ipoints,
                InputArray cameraMatrix, InputArray distCoeffs,
                OutputArray rvec, OutputArray tvec, bool useExtrinsicGuess, int flags )
