+++ modules/imgproc/misc/java/gen_dict.json
@@ -105,26 +105,7 @@
         }
     },
     "func_arg_fix" : {
-        "minEnclosingCircle" : { "points" : {"ctype" : "vector_Point2f"} },
-        "fitEllipse" : { "points" : {"ctype" : "vector_Point2f"} },
-        "fillPoly"   : { "pts" : {"ctype" : "vector_vector_Point"} },
-        "polylines"  : { "pts" : {"ctype" : "vector_vector_Point"} },
-        "fillConvexPoly" : { "points" : {"ctype" : "vector_Point"} },
-        "approxPolyDP" : { "curve"       : {"ctype" : "vector_Point2f"},
-                           "approxCurve" : {"ctype" : "vector_Point2f"} },
-        "arcLength" : { "curve" : {"ctype" : "vector_Point2f"} },
-        "pointPolygonTest" : { "contour" : {"ctype" : "vector_Point2f"} },
-        "minAreaRect" : { "points" : {"ctype" : "vector_Point2f"} },
-        "getAffineTransform" : { "src" : {"ctype" : "vector_Point2f"},
-                                 "dst" : {"ctype" : "vector_Point2f"} },
         "drawContours" : {"contours" : {"ctype" : "vector_vector_Point"} },
-        "findContours" : {"contours" : {"ctype" : "vector_vector_Point"} },
-        "convexityDefects" : { "contour"          : {"ctype" : "vector_Point"},
-                               "convexhull"       : {"ctype" : "vector_int"},
-                               "convexityDefects" : {"ctype" : "vector_Vec4i"} },
-        "isContourConvex" : { "contour" : {"ctype" : "vector_Point"} },
-        "convexHull" : { "points" : {"ctype" : "vector_Point"},
-                         "hull"   : {"ctype" : "vector_int"},
-                         "returnPoints" : {"ctype" : ""} }
+        "findContours" : {"contours" : {"ctype" : "vector_vector_Point"} }
     }
 }
