+++ modules/imgproc/misc/js/gen_dict.json
@@ -32,6 +32,7 @@
             "distanceTransformWithLabels",
             "drawContours",
             "drawMarker",
+            "drawFrameAxes",
             "ellipse",
             "ellipse2Poly",
             "equalizeHist",
@@ -42,26 +43,21 @@
             "findContours",
             "findContoursLinkRuns",
             "floodFill",
+            "fisheye_initUndistortRectifyMap",
             "GaussianBlur",
-            "getAffineTransform",
             "getFontScaleFromHeight",
-            "getPerspectiveTransform",
             "getRectSubPix",
-            "getRotationMatrix2D",
             "getStructuringElement",
             "grabCut",
             "HoughCircles",
             "HoughLines",
             "HoughLinesP",
-            "HuMoments",
             "integral",
             "integral2",
-            "invertAffineTransform",
             "Laplacian",
             "line",
             "matchTemplate",
             "medianBlur",
-            "moments",
             "morphologyEx",
             "polylines",
             "preCornerDetect",
@@ -71,7 +67,6 @@
             "rectangle",
             "remap",
             "resize",
-            "rotatedRectangleIntersection",
             "Scharr",
             "sepFilter2D",
             "Sobel",
