+++ modules/imgproc/src/lsd.cpp
