+++ modules/imgproc/src/precomp.hpp
@@ -44,6 +44,7 @@
 #define __OPENCV_PRECOMP_H__
 
 #include "opencv2/imgproc.hpp"
+#include "opencv2/geometry.hpp"
 #include "opencv2/core/utility.hpp"
 
 #include "opencv2/core/private.hpp"
