+++ modules/imgproc/test/test_contours.cpp
@@ -219,20 +219,5 @@ TEST(Imgproc_DrawContours, MatListOfMatIntScalarInt)
     EXPECT_EQ(nz, 0);
 }
 
-TEST(Imgproc_Moments, degenerateContours)
-{
-    std::vector<cv::Point> c1;
-    c1.push_back(cv::Point(10,10));
-    cv::Moments m1 = cv::moments(c1, false);
-    EXPECT_EQ(m1.m00, 0);
-
-    std::vector<cv::Point> c2;
-    c2.push_back(cv::Point(0,0));
-    c2.push_back(cv::Point(5,5));
-    c2.push_back(cv::Point(10,10));
-    cv::Moments m2 = cv::moments(c2, false);
-    EXPECT_EQ(m2.m00, 0);
-}
-
 }} // namespace
 /* End of file. */
