+++ modules/imgproc/test/test_drawing.cpp
