+++ modules/imgproc/test/test_precomp.hpp
@@ -7,8 +7,9 @@
 #include "opencv2/ts.hpp"
 #include "opencv2/ts/ts_gtest.h"
 #include "opencv2/ts/ocl_test.hpp"
-#include "opencv2/imgproc.hpp"
 #include "opencv2/core.hpp"
+#include "opencv2/imgproc.hpp"
+#include "opencv2/geometry.hpp"
 
 #include "opencv2/core/private.hpp"
