+++ samples/cpp/snippets/distrans.cpp
@@ -1,4 +1,5 @@
 #include <opencv2/core/utility.hpp>
+#include "opencv2/geometry.hpp"
 #include "opencv2/imgproc.hpp"
 #include "opencv2/imgcodecs.hpp"
 #include "opencv2/highgui.hpp"
