+++ samples/cpp/snippets/warpPerspective_demo.cpp
@@ -5,6 +5,7 @@
 @modified by Suleyman TURKMEN
 */
 
+#include "opencv2/geometry.hpp"
 #include "opencv2/imgproc.hpp"
 #include "opencv2/imgcodecs.hpp"
 #include "opencv2/highgui.hpp"
