+++ samples/cpp/tutorial_code/ImgTrans/Geometric_Transforms_Demo.cpp
@@ -6,6 +6,7 @@
 
 #include "opencv2/imgcodecs.hpp"
 #include "opencv2/highgui.hpp"
+#include "opencv2/geometry.hpp"
 #include "opencv2/imgproc.hpp"
 #include <iostream>
