+++ samples/cpp/tutorial_code/ImgTrans/imageSegmentation.cpp
@@ -4,6 +4,7 @@
  */
 
 #include <opencv2/core.hpp>
+#include <opencv2/geometry.hpp>
 #include <opencv2/imgproc.hpp>
 #include <opencv2/highgui.hpp>
 #include <iostream>
