+++ samples/cpp/tutorial_code/objectDetection/detect_board.cpp
@@ -1,5 +1,6 @@
 #include <iostream>
 #include <vector>
+#include <opencv2/imgproc.hpp>
 #include <opencv2/highgui.hpp>
 #include <opencv2/objdetect/aruco_detector.hpp>
 #include "aruco_samples_utility.hpp"
