+++ samples/cpp/tutorial_code/objectDetection/detect_board_charuco.cpp
@@ -1,4 +1,5 @@
 #include <opencv2/highgui.hpp>
+#include <opencv2/imgproc.hpp>
 //! [charucohdr]
 #include <opencv2/objdetect/charuco_detector.hpp>
 //! [charucohdr]
