+++ samples/cpp/tutorial_code/objectDetection/detect_markers.cpp
@@ -1,3 +1,4 @@
+#include <opencv2/imgproc.hpp>
 #include <opencv2/highgui.hpp>
 #include <opencv2/objdetect/aruco_detector.hpp>
 #include <iostream>
