+++ samples/dnn/text_detection.cpp
@@ -27,6 +27,7 @@
 #include <iostream>
 #include <fstream>
 
+#include <opencv2/geometry.hpp>
 #include <opencv2/imgproc.hpp>
 #include <opencv2/highgui.hpp>
 #include <opencv2/dnn.hpp>
